Rename from utility to validator_utility
