+++ lib/fcoin/validator/validator_utility.rb
